ENH: Add wider instead of narrower than normal stitching triangle strip
When determining the evaluated linear space, it was done so that the end of the linear space was whatever was still not touching the starting point (in the evaluated space). This resulted in possibly a very narrow triangle/quad strip, which was suboptimal in certain cases such as FEM analysis. Now an extra sample is subtracted from the linear space, so instead of narrower than the normal, wider than the normal elements are inserted at stitching.