BUG: Fix wrap around issue with iterative linear space search
In certain cases, especially with large number of control points the linear space was such that the initial endpoint defined for the iterative search was beyond the point searched. Worked around be increasing the parameter for the initial endpoint. This results a lightly slower but more reliable linear space search when wrapping around.